feat: change question id field type to long

+++ scrapper/src/main/java/edu/java/clients/stackoverflow/StackOverflowClient.java
@@ -1,5 +1,5 @@
 package edu.java.clients.stackoverflow;
 
 public interface StackOverflowClient {
-    QuestionResponse fetchQuestion(Integer id);
+    QuestionResponse fetchQuestion(long id);
 }
